requirements.txt
pillow 
pytesseract 
pyautogui
pyscreenshot